[148] Move test file and cleanup
